toaster: Fix breadcrumb in Chrome
Breadcrumb list items were wrapping because the content attribute was set
to an empty string that added a space. The correct way to override the
content property is setting it to 'none'. Removing the extra space
required adjustments to the breadcrumb padding.

The commit also does a bit of reorg in the css declarations.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>