toaster: set giturl on BitbakeVersion objects
In non-poky environments, the giturl was being set to "" on
BitbakeVersion objects when loadconf.py ran. This resulted
in errors when trying to checkout bitbake, as the giturl was
not set correctly.

Ensure that a giturl is set when the giturl is set to "" by
treating it the same as if it was set to None.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>