toaster: correctly revert to default sort when column is hidden
If a column is hidden while being used as the sort column for
a ToasterTable, and the default sort column for the table uses
a reverse sort (e.g. -completed_on), we don't catch the correspondence
between the default orderby for the table and this column. The
result is that the sort stays on the hidden column.

Fix this by making sure that we consider default_orderby params
which have a leading "-".

Also fix how we determine the default sort order for a table, so that
when the sort is reverted, and the table has a default reverse sort
like "-completed_on", reapply that reverse sort. (Previously, we
sorted by the column, but only in ascending order.)

[YOCTO #9011]

Signed-off-by: Elliot Smith <elliot.smith@intel.com>