toasterui: use BuildInit to capture builds with bad targets
If a build has bad targets which cause bitbake to fail before
BuildStarted is fired, we are unable to represent this in Toaster,
as the build was not detected at all.

Use the BuildInit event so we can capture these failures.

BuildStarted is retained as the event which causes a build object
to be created, to support older versions of bitbake which don't
have BuildInit.

[YOCTO #8440]

Signed-off-by: Elliot Smith <elliot.smith@intel.com>