Filter POST requests by submitter
Instead of a simple boolean parameter to turn on or off POSTing,
change the 'post' argument to 'include-post-users' which is string
indicating a regex; if any of the items's submitter match the regex,
then the POST request is done. By default the regex is empty, so
no posting is done.

This patch just supports patchwork series, not local mboxes.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>