wayland-protocols: Add recipe
wayland-protocols is a collection of Wayland protocols split
out from weston. It is a build dependency for Weston 1.10.

Protocols are not arch dependent but the recipe is not allarch because
A) getting pkgconfig paths correct with multilib+allarch is very tricky
B) multiple builds are not costly (it's just copying plain xml files)

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>