tcl: Only set BINCONFIG_GLOB for target builds
The recent change of how ${bindir_crossscripts} is installed to the
sysroot had the unforeseen effect that tclConfig.sh in the sysroot
contained invalid paths, which caused postgresql to fail to build.
The change in the contents of tclConfig.sh was due to that it was
previously installed to the sysroot after binconfig had done its work,
and was thus unaffected by it.

This change makes sure the contents of tclConfig.sh is the same as
before, both in the sysroots and on target. This will make postgresql
build again.

This also changes the BINCONFIG_GLOB to only match tclConfig.sh.
Before it would also match itclConfig.sh, tclooConfig.sh and
tdbcConfig.sh, which were consequently installed to the sysroots.
However, that seems to have been accidentally introduced with the use
of binconfig (based on what is installed in do_install()).

Signed-off-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>