matchbox-session-sato: Add meta-theme index file
This file used to be provided by gtk-sato-engine and is used by
matchbox-config-gtk only (to be able to show a single selection for
Gtk+, icon and Matchbox themes). This is a more logical place for it.

Also explicitly set the matchbox theme gconf key (even if it's the
default already).

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>