matchbox-desktop: Add Sato folder configuration
matchbox-desktop folder configuration for Sato was in a separate
project 'matchbox-desktop-sato'. Skip the extra overhead: remove
matchbox-desktop-sato and include the files here.

Drop Utilities folder and include those icons in the Applictions
folder in an effort to make the common case a little more useful:
Now e.g. Terminal icon should be visible on startup.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>