packagegroup-self-hosted: port to Python 3
A self-hosting packagegroup needs to ship Python 3 for bitbake, not Python 2.

Remove python-compiler as it was a standard module in Py2 and removed in Py3.

Remove python-misc as it is pulled in via python3-modules.

Remove python-rpm as it is only required by Smart, and if smart is installed
then python-rpm will be too.

[ YOCTO #9713 ]

(From OE-Core rev: ee627d7a833e0d85f691729c4d2f774516a8f425)

Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>