toaster: build data Add css for highlight animation
When you click on a file name that is installed in a root files system,
or in a task order number, Toaster highlights the relevant table row, so
that you know which one is the one you selected.

This patch sets the required css classes for the highlight to work.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>