toaster-manual: change mysql driver
Changed mysql driver to mysqlclient as MySQLdb doesn't
support Python 3 yet.

Here is a quote from Django 1.9 documentation regarding this:
mysqlclient is a fork of MySQLdb which notably supports
Python 3 and can be used as a drop-in replacement for
MySQLdb. At the time of this writing, this is the
recommended choice for using MySQL with Django.

Signed-off-by: Ed Bartosh <ed.bartosh@linux.intel.com>