fetch2/git.py: Sanitize the url to remove trailing '/' from the path [WIP]
A git url should not end with a trailing slash ('/').  Inside of the git tool
itself there are numerous commits that strip any trailing slashes from the
URI.  This patch ensures that behavior is matched inside of the fetcher.

This is important because of the mirroring code, '/' is replace with '.'.  So
two equivalent mirrors: example.org/git/example and example.org/git/example/
will show up as different MIRRORNAMEs.  In addition, the path the files are
cloned into is also affected.  Stripping the trailing slash ensures both
equivalent (to git) URIs are treated the same.

In addition, some filesystems, such as Windows, do not permit a directory
name to end with a '.' character.  Typically this is not an issue, but we
have users that want to transfer or share mirrored downloads via windows
platforms and shares.

(based on initial work from Jason Wessel and Robert Yang.)

Note: this current fails to pass the selftest due to the URI changing
in the fetch.py's 'test_uri' test case.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>