devtool: update-recipe: find and use existing files directories
devtool update-recipe was defaulting to the ${BPN} named directory when
adding patches, but that meant if you already had files in a ${BP} named
directory (i.e. name and version) or "files" then you'd end up with two
directories, which is usually not what you want. Look through FILESPATH
and take the first one that's the same level or one level down from the
recipe and already exists, if any.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>