site: Move rsync_cv_HAVE_C99_VSNPRINTF to common-linux
If we do not cache is then configure in distcc will try
to compute it using AC_TRY_RUN which will give some result
on x86-64 host when target is x86-64 but it generally is wrong
for cross compilation. We therefore cache it for all linux
targets.

This issue is highlighted by clang when compiling distcc
where it figures that its cross compiling and therefore triggers
it to include its own implementation of snprintf() which does
not go well with clang, gcc compiles it fine thats why we
never saw the problem thus far.

(From OE-Core rev: e940dfcb5ad4017e5fe616c583253439603656db)

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>