patchtest: Do not apply/merge patch by default
Most of the tests (in fact, all except bitbake/parse and python/lint) do not
need the patch to be merged, so it makes sense to default it to 'no merge'.
Also touching the test-* scripts, which make use of the related flags. Use
the verb 'merge' instead of 'apply'.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>