toaster: import layer Hide the layer directory form
In the import layer form, we are currently not checking which radio
button is selected and hiding the form controls accordingly on page
load: we only do this when a change happens to the radio buttons.

With this approach, we must set the display property of the layer
directory form controls to 'none', so that they are hidden from
view when the page loads with the 'Git' option selected.

Signed-off-by: Belen Barros Pena <belen.barros.pena@linux.intel.com>