toaster: fix the unidentified layer for local imported layers
When local imported layers are imported there is an issue
noticed in orm_layers table. The "Unidentified layer" row
gets added to the database. So this change helps us to
do a check if the path is a locally imported one, if so
then we cannot consider it as unidentified layer.

[YOCO #9911]

Signed-off-by: Sujith H <sujith.h@gmail.com>