buildinfohelper: prevent creation of bogus cli builds
The buildinfohelper.close() method is called for both
CommandFailed and BuildCompleted events. In cases where both of
these events occur, e.g. when cancelling a build, this means
that close() is called twice.

Previously, we assumed that close() would only be called once
and that it would reset the internal state, removing the build
in progress. But when buildinfohelper was reworked to use an
_ensure_build() method at the top of all methods which needed
the build in the internal state, we were calling _ensure_build()
twice (via buildinfohelper.update_build_information()): once
when BuildCompleted fired, and again when CommandFailed fired.

The result of this was that we were correctly updating
buildinfohelper's internal state, then removing the build from
the internal state on getting BuildCompleted. But we were then
creating a new, empty Build object when CommandFailed fired,
which looked like a "blank" cli build in the UI.

By ensuring that buildinfohelper.update_build_information()
only updates any existing Build object in internal state
(and doesn't create a Build if there isn't one), we ensure that
two calls to buildinfohelper.update_build_information() don't
create a bogus cli build.

Signed-off-by: Elliot Smith <elliot.smith@intel.com>