ref-manual: Updated the section on viewing dependencies
Fixes [YOCTO #10131]

The section was renamed "Viewing Dependencies" for consistency.
The section was moved up to be the third item in the sub-section
list.  The section was extensively re-written to provide more
clarity and options for the user to view dependencies.

(From yocto-docs rev: d521c3aabe6ded105cde6f7b3563c85340f759fd)

Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>