runqemu: add OTA image
Tipical use of OSTree deployment system relies on u-boot and specific
boot filesysem organization. It would be useful to be able to emulate
such system natively with qemu. On x86 machine u-boot can be run either
as BIOS or with Coreboot BIOS, default SeaBIOS is not supported.
Respective changes are made to runqemu.

Signed-off-by: Anton Gerasimov <anton@advancedtelematic.com>