gettext_0.16.1: use musl gettext implementation
gettext uses internal symbols to detect whether the
implementation is compatible with GNU gettext. However,
these symbols don't are not part of the public API, they
are specific to glibc.

While musl implements the GNU gettext *API* version 1 and 2
  http://www.openwall.com/lists/musl/2015/04/16/3
it doesn't implement glibc internals. This means that
gettext fails to detect musl's working implementation.

More recent versions of gettext have changed the way
GNU gettext compatibility is done
  https://lists.gnu.org/archive/html/bug-gettext/2016-04/msg00000.html
  http://git.savannah.gnu.org/cgit/gettext.git/commit/gettext-runtime/m4/gettext.m4?id=b67399b40bc5bf3165b09e6a095ec941d4b30a97
and while we could backport the corresponding patch to
gettext.m4, we avoid doing that so as to avoid any
potential GPL-v3 issues.

So instead we force ./configure to assume that the gettext
implementation of the c-library (musl) is compatible.

As a side-effect, this also reduces image sizes as the
internal gettext implementation isn't built anymore, and
it's otherwise packaged into the main gettext package
which blows up the image as the main gettext package
contains a lot of things.

Similarly, libintl.h isn't generated anymore, as the one
from musl is OK.

(From OE-Core rev: 948f0bd162f0b1b0375db884e99a2338f47e8527)

Signed-off-by: André Draszik <git@andred.net>
Signed-off-by: Ross Burton <ross.burton@intel.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>