poky-tiny: Add options to select binutils 2.24 and enable LTO
Add a commented-out BINUVERSION that points to Linux binutils
2.24.51.0.3, and which currently should only be uncommented when
building the kernel i.e. build the image first using the normal
binutils, then uncomment this BINUVERSION and build the kernel.  The
recipe for Linux binutils 2.24.51.0.3 is in meta-intel.

Also add lines enabling lto and lto-debug for the kernel.  Uncomment
those as well.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>