scripts: include the 'since' parameter
This parameter indicates a starting point in time to poll events. Thus in case
user does not provide series.revisions, a timestamp can be provided or if not
of these are provided, it will rely on the .git-poll file. If the latter is not
found, the system will try to poll ALL events, so next fetch the system is ready.

Signed-off-by: Leonardo Sandoval <leonardo.sandoval.gonzalez@linux.intel.com>